Prepare for localization: get rid of fa:ImageAwesome, introduce lang dll
