+++ Up2dateService/Up2dateConsole/Helpers/EnumToVisibilityConverter.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Data;
+
+namespace Up2dateConsole.Helpers
+{
+
+    public class EnumToVisibilityConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo language)
+        {
+            if (value == null || parameter == null || !(value is Enum))
+                return Visibility.Collapsed;
+
+            var currentState = value.ToString();
+            var stateStrings = parameter.ToString();
+            var found = false;
+
+            foreach (var state in stateStrings.Split(','))
+            {
+                found = currentState == state.Trim();
+                if (found) break;
+            }
+
+            return found ? Visibility.Visible : Visibility.Collapsed;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
