+++ Up2dateService/Up2dateConsole/StateIndicator/ServiceState.cs
@@ -1,4 +1,4 @@
-﻿namespace Up2dateConsole
+﻿namespace Up2dateConsole.StateIndicator
 {
     public enum ServiceState
     {
