+++ Up2dateService/Up2dateConsole/Styles.xaml
@@ -1,0 +1,46 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:clr="clr-namespace:System;assembly=mscorlib"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    
+    <clr:String x:Key="KeyIcon">&#xf084;</clr:String>
+    <clr:String x:Key="RefreshIcon">&#xf021;</clr:String>
+    <clr:String x:Key="ExclamationTriangleIcon">&#xf071;</clr:String>
+    <clr:String x:Key="CheckCircleIcon">&#xf058;</clr:String>
+    <clr:String x:Key="QuestionIcon">&#xf128;</clr:String>
+    <clr:String x:Key="SpinnerIcon">&#xf110;</clr:String>
+
+    <Style x:Key="AwesomeIcon" TargetType="{x:Type TextBlock}">
+        <Setter Property="FontFamily" Value="{StaticResource FontAwesome}"/>
+        <Setter Property="VerticalAlignment" Value="Center"/>
+        <Setter Property="HorizontalAlignment" Value="Center"/>
+    </Style>
+
+    <Style x:Key="ButtonIcon" TargetType="{x:Type TextBlock}" BasedOn="{StaticResource AwesomeIcon}">
+        <Setter Property="Foreground" Value="Gray"/>
+        <Setter Property="FontSize" Value="16"/>
+    </Style>
+
+    <Style x:Key="StateIcon" TargetType="{x:Type TextBlock}" BasedOn="{StaticResource AwesomeIcon}">
+        <Setter Property="FontSize" Value="18"/>
+    </Style>
+
+    <Style x:Key="RotatingStateIcon" BasedOn="{StaticResource StateIcon}" TargetType="{x:Type TextBlock}">
+        <Setter Property="RenderTransformOrigin" Value="0.5,0.5"/>
+        <Setter Property="RenderTransform">
+            <Setter.Value>
+                <RotateTransform/>
+            </Setter.Value>
+        </Setter>
+        <Style.Triggers>
+            <EventTrigger RoutedEvent="TextBlock.Loaded">
+                <BeginStoryboard>
+                    <Storyboard>
+                        <DoubleAnimation Duration="0:0:1" From="0" To="360" RepeatBehavior="Forever"
+                                            Storyboard.TargetProperty="(TextBlock.RenderTransform).(RotateTransform.Angle)"/>
+                    </Storyboard>
+                </BeginStoryboard>
+            </EventTrigger>
+        </Style.Triggers>
+    </Style>
+
+</ResourceDictionary>
