+++ Up2dateService/Up2dateConsole/Up2dateConsole.csproj
@@ -2,6 +2,9 @@
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
+    <UICulture>en-US</UICulture>
+  </PropertyGroup>
+  <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{06A754B8-B045-40FE-BADB-2F0469BC43BA}</ProjectGuid>
@@ -73,13 +76,17 @@
       <DependentUpon>RequestCertificateDialog.xaml</DependentUpon>
     </Compile>
     <Compile Include="Helpers\CommandLineHelper.cs" />
+    <Compile Include="Helpers\EnumToVisibilityConverter.cs" />
     <Compile Include="Helpers\SingleInstanceHelper.cs" />
     <Compile Include="IWcfClientFactory.cs" />
+    <Compile Include="StateIndicator\StateIndicatorViewModel.cs" />
+    <Compile Include="StateIndicator\StateIndicator.xaml.cs">
+      <DependentUpon>StateIndicator.xaml</DependentUpon>
+    </Compile>
     <Compile Include="ViewService\IViewService.cs" />
     <Compile Include="Helpers\ThreadHelper.cs" />
     <Compile Include="PackageItem.cs" />
-    <Compile Include="ServiceState.cs" />
-    <Compile Include="StateIndicatorViewModel.cs" />
+    <Compile Include="StateIndicator\ServiceState.cs" />
     <Compile Include="ViewService\DialogViewModelBase.cs" />
     <Compile Include="ViewService\ViewService.cs" />
     <Compile Include="WcfClientFactory.cs" />
@@ -101,6 +108,14 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
+    <Page Include="StateIndicator\StateIndicator.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Styles.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MainWindowViewModel.cs" />
@@ -149,6 +164,7 @@
     <None Include="Connected Services\ServiceReference\Up2dateShared.xsd">
       <SubType>Designer</SubType>
     </None>
+    <Resource Include="Fonts\fontawesome-webfont.ttf" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
@@ -161,9 +177,6 @@
     <WCFMetadata Include="Connected Services\" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="FontAwesome.WPF">
-      <Version>4.7.0.9</Version>
-    </PackageReference>
     <PackageReference Include="Hardcodet.NotifyIcon.Wpf">
       <Version>1.1.0</Version>
     </PackageReference>
@@ -200,5 +213,6 @@
       <LastGenOutput>Reference.cs</LastGenOutput>
     </None>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
